хитбокс игрока не коллайдит с terrain; чистка

+++ impl-demo/minecraft/src/main/kotlin/ru/cherryngine/impl/demo/minecraft/output/MinecraftPlayerMover.kt
@@ -29,8 +29,8 @@ class MinecraftPlayerMover(
                 0,
                 position.minestomVec(),
                 Vec.ZERO,
-                yawPitch.yaw.toFloat(),
-                yawPitch.pitch.toFloat(),
+                yawPitch.yaw,
+                yawPitch.pitch,
                 0
             )
         )
